Update state in app
